+++ src/backend/admin/features/adminPermission/GetServerConfig.php
@@ -8,22 +8,25 @@ use backend\Configs;
 class GetServerConfig extends HasAdminPermission {
 	
 	function exec(): array {
-		$serverSettings = Configs::getAll();
-		$serverSettings['impressum'] = [];
-		$serverSettings['privacyPolicy'] = [];
-		
 		$langCodes = Configs::get('langCodes');
-		$langCodes[] = '_';
+		$serverName = Configs::get('serverName');
+		$configs = [
+			'langCodes' => $langCodes,
+			'defaultLang' => Configs::get('defaultLang')
+		];
+		$translationData = [];
+		
 		$serverStore = Configs::getDataStore()->getServerStore();
 		foreach($langCodes as $code) {
-			$impressum = $serverStore->getImpressum($code);
-			if(!empty($impressum))
-				$serverSettings['impressum'][$code] = $impressum;
-			
-			$privacyPolicy = $serverStore->getPrivacyPolicy($code);
-			if(!empty($impressum))
-				$serverSettings['privacyPolicy'][$code] = $privacyPolicy;
+			$translationData[$code] = [
+				'serverName' => $serverName[$code],
+				'impressum' => $serverStore->getImpressum($code),
+				'privacyPolicy' => $serverStore->getPrivacyPolicy($code)
+			];
 		}
-		return $serverSettings;
+		return [
+			'configs' => $configs,
+			'translationData' => $translationData
+		];
 	}
 }