+++ src/frontend/js/helpers/knockout_own_mapping.js
@@ -1,5 +1,29 @@
-import ko from "knockout"
+import ko from "knockout";
 import {Studies} from "../main_classes/studies";
+
+
+let copyArray = function(a) {
+	let r = [];
+	for(let i=0, max=a.length; i<max; ++i) {
+		let value = a[i];
+		r.push(typeof value === "function" ? value() : value);
+	}
+	return r;
+}
+let compareArray = function(a, originalArray) {
+	if(originalArray.length !== a.length)
+		return true;
+	for(let i=0, max=originalArray.length; i<max; ++i) {
+		let value = a[i];
+		if(originalArray[i] !== (typeof value === "function" ? value() : value))
+			return true;
+	}
+	return false;
+}
+let _uniqueIdCounter = 100;
+let getUniqueId = function() {
+	return ++_uniqueIdCounter;
+}
 
 export const OwnMapping = {
 	indexOf: function(obj, s) {
@@ -9,138 +33,29 @@ export const OwnMapping = {
 		}
 		return -1;
 	},
-	_observableArray: function(a) {
-		let copyArray = function() {
-			let r = [];
-			for(let i=0, max=a.length; i<max; ++i) {
-				let value = a[i];
-				r.push(typeof value === "function" ? value() : value);
-			}
-			return r;
-		}
-		let compareArray = function() {
-			if(originalArray.length !== a.length)
-				return true;
-			for(let i=0, max=originalArray.length; i<max; ++i) {
-				let value = a[i];
-				if(originalArray[i] !== (typeof value === "function" ? value() : value))
-					return true;
-			}
-			return false;
-		}
-		
+	_observableArray: function(a, defaultValue, translationId) {
 		let obs = ko.observableArray(a);
-		let originalArray = copyArray();
-		obs.___defaultValue = [];
+		let originalArray = copyArray(a);
+		obs.___defaultValue = defaultValue || [];
 		obs.___isDirty = function() {
-			return compareArray();
+			return compareArray(a, originalArray);
 		}
 		obs.___unsetDirty = function() {
-			originalArray = copyArray();
+			originalArray = copyArray(a);
 		};
+		if(translationId)
+			obs.___translationId = translationId;
 		let self = this;
 		obs.indexOf = function(s) {return self.indexOf(this(), s);};
 		return obs;
 	},
 	
-	_translatableObservableValue: function(value, langCode, defaultValue) {
-		let self = this;
-		let isArray = Array.isArray(defaultValue);
-		let data;
-		if(isArray)
-			data = {_: this.fromJS(value.hasOwnProperty(langCode) ? value[langCode] : defaultValue[langCode], defaultValue)};
-		else
-			data = {_: ko.observable(value.hasOwnProperty(langCode) ? value[langCode] : defaultValue)};
-		
-		let changedValues = {};
-		let thisObj = ko.pureComputed({
-			read: function() {
-				let code = Studies.tools ? Studies.tools.currentLang() : "_";
-				
-				if(!data.hasOwnProperty(code))
-					//we can not just pass data._ because this observable is used for update events
-					// meaning that when write creates a new observable, nobody would be notified about the change
-					data[code] = ko.observable(data._());
-				
-				return data[code]();
-			},
-			write: function(newValue) {
-				let code = Studies.tools ? Studies.tools.currentLang() : "_";
-				if(!data.hasOwnProperty(code)) //very unlikely to happen since it is already created in read
-					data[code] = ko.observable(newValue);
-				
-				let value = data[code];
-				if(!changedValues.hasOwnProperty(code))
-					changedValues[code] = value();
-				else if(changedValues[code] === newValue)
-					delete changedValues[code];
-				
-				value(newValue);
-			}
-		});
-		thisObj.___defaultValue = defaultValue;
-		thisObj.___isDirty = function() {
-			for(let k in changedValues) {
-				if(changedValues.hasOwnProperty(k))
-					return true;
-			}
-			return false;
-		};
-		thisObj.___unsetDirty = function() {
-			changedValues = {};
-		};
-		thisObj.___setLang = function(code, newValue) {
-			if(!data.hasOwnProperty(code)) {
-				if(isArray)
-					data[code] = self.fromJS(newValue);
-				else
-					data[code] = ko.observable(newValue);
-			}
-			else
-				data[code](newValue);
-			
-			if(Studies.tools && Studies.tools.currentLang() === code) { //the new language needs to be displayed right now
-				//the lang was not there before, you _() was used instead. So we tell knockout that _ has changed to force a reread:
-				data._.valueHasMutated();
-			}
-		};
-		thisObj.___getLang = function(code) {
-			// return data.hasOwnProperty(code) ? data[code]() : thisObj.___defaultValue;
-			return data.hasOwnProperty(code) ? data[code]() : data._();
-		};
-		
-		if(isArray) { //it is not an observableArray. So we "fake" it by adding all necessary methods:
-			let getObs = function() {
-				let code = Studies.tools ? Studies.tools.currentLang() : "_";
-				return data.hasOwnProperty(code) ? data[code] : data._;
-			}
-			
-			//for a list of methods see: https://knockoutjs.com/documentation/observableArrays.html
-			//"remove" breaks things for some reason. But since we dont use it anyway we just dont include it
-			let methods = ["valueHasMutated", "slice", "push", "pop", "unshift", "shift", "reverse", "sort", "splice", "sorted", "removeAll", "destroy", "destroyAll"];
-			methods.forEach(function(fu) {
-				thisObj[fu] = function() {
-					let obj = getObs();
-					obj[fu].apply(obj, arguments);
-					
-					let code = Studies.tools ? Studies.tools.currentLang() : "_";
-					if(!changedValues.hasOwnProperty(code))
-						changedValues[code] = true;
-					else if(!obj.___isDirty())
-						delete changedValues[code];
-				};
-			});
-			thisObj.indexOf =  function(s) {return self.indexOf(getObs()(), s);};
-		}
-		
-		return thisObj;
-	},
-	_observableValue: function(value, defaultValue) {
+	_observableValue: function(value, defaultValue, translationId) {
 		let obs = ko.observable(value);
-		this._toOwnObservableValue(obs, value, defaultValue);
+		this._toOwnObservableValue(obs, value, defaultValue, translationId);
 		return obs;
 	},
-	_toOwnObservableValue: function(obs, value, defaultValue) {
+	_toOwnObservableValue: function(obs, value, defaultValue, translationId) {
 		if(!obs.hasOwnProperty("___defaultValue")) {
 			obs.___defaultValue = defaultValue;
 			obs.___isDirty = function() {
@@ -149,6 +64,8 @@ export const OwnMapping = {
 			obs.___unsetDirty = function() {
 				value = obs();
 			}
+			if(translationId)
+				obs.___translationId = translationId;
 		}
 	},
 	fromJS: function(value, defaultObj) {
@@ -161,82 +78,167 @@ export const OwnMapping = {
 		else
 			type = typeof value;
 		
-		if(type === "object") {
-			if(Array.isArray(value)) {
-				let a = [];
-				// let newDefaultObj = defaultObj ? defaultObj.$ : undefined;
-				
-				//Note: when defaultObj is an array, it is only meant to replace value in case value is undefined. Here it is unneeded
-				// If not then value has object entries that need to be checked further against defaultObj
-				let newDefaultObj = Array.isArray(defaultObj) ? undefined : defaultObj;
-				for(let i=0, max=value.length; i<max; ++i) {
-					a.push(this.fromJS(value[i], newDefaultObj));
-				}
-				return this._observableArray(a);
+		
+		// non iterable:
+		if(type !== "object")
+			return this._observableValue(value, defaultObj);
+		
+		// array:
+		else if(Array.isArray(value)) {
+			let a = [];
+			for(let i=0, max=value.length; i<max; ++i) {
+				//its entries should only have a default if it is a container:
+				a.push(this.fromJS(value[i], (defaultObj && (typeof (defaultObj[i])) === "object") ? defaultObj[i] : undefined));
 			}
-			else if(defaultObj) {
-				let r = {};
-				for(let key in defaultObj) {
-					if(!defaultObj.hasOwnProperty(key))
-						continue;
-					
-					if(key === "$") {
-						if(defaultObj.$.hasOwnProperty("children")) {
-							let children = defaultObj.$.children;
-							for(let realKey in children) {
-								if(!children.hasOwnProperty(realKey))
-									continue;
-								
-								if(value.hasOwnProperty(realKey))
-									r[realKey] = this.fromJS(value[realKey], children[realKey]);
-								else
-									r[realKey] = this._observableArray([], children[realKey]);
-							}
-						}
-						if(defaultObj.$.hasOwnProperty("translated")) {
-							let defaultTranslations = defaultObj.$.translated;
-							for(let realKey in defaultTranslations) {
-								if(!defaultTranslations.hasOwnProperty(realKey))
-									continue;
-								
-								r[realKey] = this._translatableObservableValue(value, realKey, defaultTranslations[realKey]);
-							}
-						}
-						if(defaultObj.$.hasOwnProperty("noDefault")) {
-							let children = defaultObj.$.noDefault;
-							for(let realKey in children) {
-								if(!children.hasOwnProperty(realKey))
-									continue;
-								
-								r[realKey] = this.fromJS(value.hasOwnProperty(realKey) ? value[realKey] : null);
-							}
-						}
-					}
-					else {
-						let newDefaultObj = defaultObj[key];
-						
-						if(value.hasOwnProperty(key))
-							r[key] = this.fromJS(value[key], newDefaultObj);
-						else
-							r[key] = this.fromJS(newDefaultObj, newDefaultObj); //the new value may have $... values from Defaults. But they will be filtered anyway
-					}
+			return this._observableArray(a, defaultObj);
+		}
+		
+		
+		let r = {};
+		
+		// object, but no default information:
+		if(!defaultObj) {
+			for(let key in value) {
+				if(value.hasOwnProperty(key)) {
+					r[key] = this.fromJS(value[key]);
 				}
-				
-				return r;
 			}
-			else {
-				let r = {};
-				for(let key in value) {
-					if(value.hasOwnProperty(key)) {
-						r[key] = this.fromJS(value[key]);
+			return r;
+		}
+		
+		//loop all normal values
+		for(let key in defaultObj) {
+			if(!defaultObj.hasOwnProperty(key) || key === "$")
+				continue;
+			
+			let newDefaultObj = defaultObj[key];
+			
+			r[key] = value.hasOwnProperty(key)
+				? this.fromJS(value[key], newDefaultObj)
+				: this.fromJS(newDefaultObj, newDefaultObj); //the new value may have $... values from Defaults. But they will be filtered anyway
+		}
+		
+		if(!defaultObj.hasOwnProperty("$"))
+			return r;
+		
+		// handle special $ value:
+		if(defaultObj.$.hasOwnProperty("children")) {
+			let children = defaultObj.$.children;
+			for(let realKey in children) {
+				if(!children.hasOwnProperty(realKey))
+					continue;
+				
+				if(value.hasOwnProperty(realKey)) {
+					//all entries need to use the same default:
+					let a = [];
+					let entries = value[realKey];
+					let newDefault = children[realKey];
+					for(let i=0, max=entries.length; i<max; ++i) {
+						a.push(this.fromJS(entries[i], newDefault));
 					}
+					r[realKey] = this._observableArray(a);
 				}
-				return r;
+				else
+					r[realKey] = this._observableArray([]);
 			}
 		}
-		else
-			return this._observableValue(value, defaultObj)
+		if(defaultObj.$.hasOwnProperty("translated")) {
+			let defaultTranslations = defaultObj.$.translated;
+			for(let realKey in defaultTranslations) {
+				if(!defaultTranslations.hasOwnProperty(realKey))
+					continue;
+				
+				let newValue = value[realKey];
+				let defaultValue = defaultTranslations[realKey];
+				if(Array.isArray(defaultValue)) {
+					let obs = this.fromJS(newValue, defaultTranslations[realKey]); // this will be an _observableArray filled with observables
+					obs.___translationId = getUniqueId();
+					r[realKey] = obs;
+				}
+				else
+					r[realKey] = this._observableValue(newValue, defaultTranslations[realKey], getUniqueId());
+			}
+		}
+		
+		return r;
 	},
+	
+	bindNewLanguageContainer: function(obs, translationJs, translations) {
+		// Assumption: objs has the same structure as translationJs
+		// So translatable values are ordered the same
+		
+		translations = translations || {};
+		let value = typeof obs === "function" ? obs() : obs;
+		if(typeof value !== "object")
+			return;
+		
+		if(Array.isArray(value)) {
+			for(let i=0, max=value.length; i<max; ++i) {
+				this.bindNewLanguageContainer(value[i], translationJs ? translationJs[i] : null, translations);
+			}
+			return;
+		}
+		
+		for(let key in value) {
+			if(!value.hasOwnProperty(key))
+				continue
+			
+			let currentObs = value[key];
+			
+			if(!currentObs.hasOwnProperty("___translationId")) {
+				this.bindNewLanguageContainer(value[key], translationJs ? translationJs[key] : null, translations);
+				continue;
+			}
+			
+			let id = currentObs.___translationId;
+			let translatedValue = translationJs && translationJs.hasOwnProperty(key)
+				? translationJs[key]
+				: currentObs.___defaultValue
+			
+			if(Array.isArray(currentObs())) {
+				let obsArray = this.fromJS(translatedValue || [], currentObs.___defaultValue || []); // this will be an _observableArray filled with observables
+				obsArray.___translationId = id;
+				translations[id] = obsArray;
+			}
+			else
+				translations[id] = this._observableValue(translatedValue || "", currentObs.___defaultValue || "", id);
+			
+			// translations[id] = Array.isArray(currentObs)
+			// 	? this._observableArray(translatedValue || [], currentObs.___defaultValue || [], id)
+			// 	: this._observableValue(translatedValue || "", currentObs.___defaultValue || "", id);
+		}
+		return translations;
+	},
+	
+	switchLanguage(obs, fromLanguageContainer, toLanguageContainer) {
+		// Rationale:
+		// loop obs. When encountering a translated entry, we:
+		// - check if its in fromLanguageContainer. If not, copy the observable over (only happens, when a new element was added or default lang was not initialized yet)
+		// - check if its in toLanguageContainer. If not, we create an empty observable (only happens, when a new element was added or lang is new)
+		// - copy the observable of target lang into obs
+		
+		let checkTranslation = function(currentObs, parent, key) {
+			if(!currentObs.hasOwnProperty("___translationId"))
+				return;
+			
+			let id = currentObs.___translationId;
+			if(fromLanguageContainer && !fromLanguageContainer.hasOwnProperty(id))
+				fromLanguageContainer[id] = currentObs;
+			
+			if(!toLanguageContainer.hasOwnProperty(id)) {
+				toLanguageContainer[id] = Array.isArray(currentObs)
+					? self._observableArray([], currentObs.___defaultValue || [], id)
+					: self._observableValue(currentObs() || "", currentObs.___defaultValue || "", id);
+			}
+			parent[key] = toLanguageContainer[id];
+			
+		}
+		
+		let self = this;
+		this.loopAll(obs, checkTranslation, checkTranslation);
+	},
+	
+	
 	
 	update: function(old_obj, new_obj, defaultObj, isDefaultArray) {
 		for(let key in new_obj) {
@@ -292,76 +294,35 @@ export const OwnMapping = {
 		//there is no need to delete unneeded properties in old_obj since they will be filtered and it will happen rarely anyway
 	},
 	
-	add_lang: function(obj, defaultObj, langObj, code) {
-		if(defaultObj.hasOwnProperty("$")) {
-			let $ = defaultObj.$;
-			if($.hasOwnProperty("translated")) {
-				let defaultTranslations = $.translated;
-				for(let key in defaultTranslations) {
-					if(!defaultTranslations.hasOwnProperty(key))
-						continue;
-					
-					let value;
-					if(langObj && langObj.hasOwnProperty(key))
-						value = langObj[key];
-					else
-						value = defaultTranslations[key];
-					obj[key].___setLang(code, value);
-				}
+	loopAll: function(obj, valueFu, arrayFu, currentParent, currentKey) {
+		let value = typeof obj === "function" ? obj() : obj;
+		
+		if(typeof value !== "object")
+			return valueFu(obj, currentParent, currentKey);
+		
+		if(Array.isArray(value)) {
+			for(let i=0, max=value.length; i<max; ++i) {
+				this.loopAll(value[i], valueFu, arrayFu, value, i);
 			}
-			
-			if($.hasOwnProperty("children")) {
-				let defaultChildren = $.children;
-				for(let key in defaultChildren) {
-					if(!defaultChildren.hasOwnProperty(key))
-						continue;
-					let array = obj[key]();
-					let langArray = (langObj && langObj.hasOwnProperty(key)) ? langObj[key] : [];
-					for(let i = array.length - 1; i >= 0; --i) {
-						this.add_lang(array[i], defaultChildren[key], langArray[i], code);
-					}
-				}
-			}
+			if(arrayFu)
+				arrayFu(obj, currentParent, currentKey);
+			return;
 		}
 		
-	},
-	
-	
-	loopAll: function(obj, valueFu, arrayFu) {
-		let value = typeof obj === "function" ? obj() : obj;
-		if(typeof value === "object") {
-			if(Array.isArray(value)) {
-				let a = [];
-				for(let i=0, max=value.length; i<max; ++i) {
-					let v = this.loopAll(value[i], valueFu, arrayFu);
-					if(v !== undefined)
-						a.push(v);
-				}
-				if(arrayFu)
-					arrayFu(obj);
-				return a;
-			}
-			else {
-				let o = {};
-				for(let key in value) {
-					if(value.hasOwnProperty(key)) {
-						let v = this.loopAll(value[key], valueFu, arrayFu);
-						if(v !== undefined)
-							o[key] = v;
-					}
-				}
-				return o;
-			}
+		for(let key in value) {
+			if(!value.hasOwnProperty(key))
+				continue;
+			
+			this.loopAll(value[key], valueFu, arrayFu, value, key);
 		}
-		else
-			return valueFu(obj);
 	},
-	subscribe: function(obj, dirtyObj, changedFu) {
+	subscribe: function(obj, changedFu) {
+		let dirtyObj = ko.observable(false);
 		let self = this;
 		let subscriptions = []; //used to keep track how many variables are dirty (and when dirty can be false again)
 		let dirtyTrack = {};
 		let subscribeFu = function(currentObj) {
-			self._toOwnObservableValue(currentObj);
+			self._toOwnObservableValue(currentObj); //in case we added new elements as ko.observable() instead of _observableValue()
 			let index = subscriptions.length;
 			subscriptions.push(currentObj.subscribe(function() {
 				if(dirtyObj) {
@@ -386,13 +347,24 @@ export const OwnMapping = {
 			}));
 		};
 		this.loopAll(obj, subscribeFu, subscribeFu);
-		return subscriptions;
+		
+		
+		return {
+			dispose: function() {
+				for(let i=subscriptions.length-1; i>=0; --i) {
+					subscriptions[i].dispose();
+				}
+			},
+			isDirty: dirtyObj
+		};
+		// return subscriptions;
 	},
 	unsetDirty: function(obj) {
 		let self = this;
 		let unsetFu = function(currentObj) {
-			self._toOwnObservableValue(currentObj);
+			self._toOwnObservableValue(currentObj); //in case we added new elements as ko.observable() instead of _observableValue()
 			currentObj.___unsetDirty();
+			currentObj.notifySubscribers();
 		};
 		this.loopAll(obj, unsetFu, unsetFu);
 	},
@@ -405,10 +377,40 @@ export const OwnMapping = {
 		});
 	},
 	toJS: function(obj) {
-		return this.loopAll(obj, function(currentObj) {
-			return (currentObj() !== currentObj.___defaultValue) ? currentObj() : undefined; //by setting it to undefined, we effectively strip the whole variable away
-		});
+		let value = typeof obj === "function" ? obj() : obj;
+		
+		if(typeof value !== "object")
+			return (value !== obj.___defaultValue) ? value : undefined; //by setting it to undefined, we effectively strip the whole variable away
+		
+		if(Array.isArray(value)) {
+			let a = [];
+			for(let i=0, max=value.length; i<max; ++i) {
+				let v = this.toJS(value[i], value, i);
+				if(v !== undefined)
+					a.push(v);
+			}
+			if(!a.length && value.length) // if all entries were default values then they were stripped while creating the array - leaving n empty array
+				return obj.___defaultValue;
+			return a;
+		}
+		
+		let o = {};
+		for(let key in value) {
+			if(!value.hasOwnProperty(key))
+				continue;
+			
+			let v = this.toJS(value[key], value, key);
+			if(v !== undefined)
+				o[key] = v;
+		}
+		return o;
 	},
+	// toJS: function(obj) {
+	// 	let sanitizeValue = function(currentObj) {
+	// 		return (currentObj() !== currentObj.___defaultValue) ? currentObj() : undefined; //by setting it to undefined, we effectively strip the whole variable away
+	// 	};
+	// 	return this.loopAll(obj, sanitizeValue);
+	// },
 	toJSON: function(obj) {
 		let a = this.toJS(obj);
 		return JSON.stringify(a);