+++ src/frontend/js/main_classes/page.js
@@ -12,6 +12,10 @@ import {NavigationRow} from "./navigation_row";
 export function Page(depth, code) {
 	let self = this;
 	
+	//used by tab-bar widget so keep tab selection after page reload
+	this.tabBarSize = 0
+	this.tabBarSelections = []
+	
 	this.isLoading = false;
 	this.codeString = code;
 	this.depth = depth;
@@ -69,7 +73,6 @@ export function Page(depth, code) {
 		title_el = createElement("div", false, {className: "title"}, {"data-bind": "text: $root.page.printTitle"}),
 		extra_el = createElement("div", false, {className: "extra"}),
 		content_el = createElement("div", "opacity: 0", {className: "page_content"}, {"data-bind": "with: dataObj"});
-		// lang_el = createElement("lang-options", false, false, {"params": "page: $root.page, data: $data"});
 	
 	top_el.appendChild(back_el);
 	top_el.appendChild(title_line_el);
@@ -222,6 +225,7 @@ export function Page(depth, code) {
 	//
 	
 	this.reload = function() {
+		self.tabBarSize = 0
 		ko.cleanNode(parentEl);
 		promise = load(pageName);
 	};