+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### ⚠️ Breaking
+- Default server language will be set to english. Please check your server settings after update
+
 ### 🚀 Added
 - Added ukrainian translations.